fix: remove raw secret resource
